# Conflicts:
#	union-app-api/src/main/java/com/union/app/api/pk/zone/收藏.java
#	union-app-api/src/main/java/com/union/app/api/pk/zone/查询单个PK.java
#	union-app-api/src/main/java/com/union/app/api/pk/审核/查询审核信息1.java
#	union-app-api/src/main/java/com/union/app/api/pk/审核/查询审核信息2.java
#	union-app-api/src/main/java/com/union/app/api/pk/审核/查询审核员审核中列表.java
#	union-app-api/src/main/java/com/union/app/api/pk/审核/查询审核员已审核列表.java
#	union-app-api/src/main/java/com/union/app/api/pk/审核/设置留言.java
#	union-app-api/target/classes/com/union/app/api/pk/zone/收藏.class
#	union-app-api/target/classes/com/union/app/api/pk/zone/查询单个PK.class
#	union-app-api/target/classes/com/union/app/api/登录和注册/用户登录加注册.class
#	union-app-common/target/classes/com/union/app/common/OSS存储/IdGenerator.class
#	union-app-common/target/classes/com/union/app/common/OSS存储/OssStorage.class
#	union-app-common/target/classes/com/union/app/common/api/registry/ApiURI.class
#	union-app-common/target/classes/com/union/app/common/api/registry/ServiceRegistry.class
#	union-app-common/target/classes/com/union/app/common/api/registry/helper/ServiceBeanHelper.class
#	union-app-common/target/classes/com/union/app/common/resultcode/ResultCodeService.class
#	union-app-common/target/classes/com/union/app/common/spring/interceptor/ControllerRequestLoggerInterceptor.class
#	union-app-common/target/classes/com/union/app/common/spring/listener/AppConfigServiceListener.class
#	union-app-common/target/classes/com/union/app/common/validate/AppValidatorService.class
#	union-app-common/target/classes/com/union/app/common/微信/AES.class
#	union-app-common/target/classes/com/union/app/common/微信/WeChatUtil.class
#	union-app-dao/src/main/java/com/union/app/dao/druid/DuridConfig.java
#	union-app-dao/target/classes/com/union/app/dao/druid/DuridConfig.class
#	union-app-dao/target/classes/com/union/app/dao/spi/AppDaoService.class
#	union-app-dao/target/classes/com/union/app/dao/spi/filter/NullFilter.class
#	union-app-domain/src/main/java/com/union/app/domain/pk/审核/ApproveUser.java
#	union-app-domain/target/classes/com/union/app/domain/pk/Post.class
#	union-app-domain/target/classes/com/union/app/domain/pk/UserCode.class
#	union-app-domain/target/classes/com/union/app/domain/房间/Room.class
#	union-app-domain/target/classes/com/union/app/domain/支付/City.class
#	union-app-domain/target/classes/com/union/app/domain/支付/Job.class
#	union-app-entity/target/classes/com/union/app/entity/order/PkOrderEntity.class
#	union-app-entity/target/classes/com/union/app/entity/pk/EmptyPostEntity.class
#	union-app-entity/target/classes/com/union/app/entity/pk/OrderEntity.class
#	union-app-entity/target/classes/com/union/app/entity/pk/PostEntity.class
#	union-app-entity/target/classes/com/union/app/entity/pk/PostStatu.class
#	union-app-entity/target/classes/com/union/app/entity/pk/UserCollectionEntity.class
#	union-app-entity/target/classes/com/union/app/entity/pk/share/ShareEntity.class
#	union-app-entity/target/classes/com/union/app/entity/pk/助力浏览评论分享/UserLikeEntity.class
#	union-app-entity/target/classes/com/union/app/entity/share/ShareFeeEntity.class
#	union-app-entity/target/classes/com/union/app/entity/应用信息表/ApplicationEntity.class
#	union-app-entity/target/classes/com/union/app/entity/标签/LabelEntity.class
#	union-app-entity/target/classes/com/union/app/entity/用户/InviteViewHistoryEntity.class
#	union-app-entity/target/classes/com/union/app/entity/用户/UserBalanceEntity.class
#	union-app-entity/target/classes/com/union/app/entity/用户/UserBalanceHistoryEntity.class
#	union-app-entity/target/classes/com/union/app/entity/用户/UserFollowEntity.class
#	union-app-entity/target/classes/com/union/app/entity/用户/UserInviteFromUserEntity.class
#	union-app-entity/target/classes/com/union/app/entity/用户/UserShareHistoryEntity.class
#	union-app-plateform/src/main/java/com/union/app/plateform/constant/常量值.java
#	union-app-plateform/target/classes/com/union/app/plateform/constant/常量值.class
#	union-app-plateform/target/classes/com/union/app/plateform/storgae/RedisConfig.class
#	union-app-plateform/target/classes/com/union/app/plateform/storgae/redis/RedisStringUtil.class
#	union-app-service/src/main/java/com/union/app/service/pk/complain/ComplainService.java
#	union-app-service/src/main/java/com/union/app/service/pk/dynamic/DynamicItem.java
#	union-app-service/src/main/java/com/union/app/service/pk/dynamic/DynamicService.java
#	union-app-service/src/main/java/com/union/app/service/pk/service/ApproveService.java
#	union-app-service/src/main/java/com/union/app/service/pk/service/PostService.java
#	union-app-service/src/main/java/com/union/app/service/pk/service/UserInfoService.java
#	union-app-service/src/main/java/com/union/app/service/quartz/TEST.java
#	union-app-service/target/classes/com/union/app/service/pk/click/ClickItem.class
#	union-app-service/target/classes/com/union/app/service/pk/click/ClickService.class
#	union-app-service/target/classes/com/union/app/service/pk/dynamic/DynamicItem.class
#	union-app-service/target/classes/com/union/app/service/pk/dynamic/DynamicService.class
#	union-app-service/target/classes/com/union/app/service/pk/service/PkService.class
#	union-app-service/target/classes/com/union/app/service/pk/service/PostService.class
#	union-app-service/target/classes/com/union/app/service/pk/service/UserInfoService.class
#	union-app-service/target/classes/com/union/app/service/quartz/TEST.class
#	union-app-start/src/main/resources/application.properties
#	union-app-util/target/classes/com/union/app/util/http/AppHttpUtil.class
#	union-app-util/target/classes/com/union/app/util/jar/JarLoader.class
#	union-app-util/target/classes/com/union/app/util/jar/JarUtil.class
#	union-app-util/target/classes/com/union/app/util/reflect/EntityAnalyser.class
#	union-app-util/target/classes/com/union/app/util/time/TimeUtils.class

+++ union-app-service/src/main/java/com/union/app/service/pk/service/PkService.java
@@ -385,7 +385,8 @@ public class PkService {
         EntityFilterChain filter = EntityFilterChain.newFilterChain(PkActiveEntity.class)
                 .compareFilter("pkId",CompareTag.Equal,pkId);
         PkActiveEntity pkActiveEntity = daoService.querySingleEntity(PkActiveEntity.class,filter);
-        daoService.deleteEntity(pkActiveEntity);
+        if(!ObjectUtils.isEmpty(pkActiveEntity)){daoService.deleteEntity(pkActiveEntity);}
+
 
     }
